fix(stories): Update component import paths after moving files
